llm-service/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
python-dotenv==1.0.1
ollama==0.1.6
chromadb==0.4.24
tiktoken==0.6.0
sentence-transformers==2.6.1
langchain==0.1.16
fastapi==0.110.2
uvicorn==0.29.0
pytest==8.1.1
mypy==1.9.0
black==24.4.0
